# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Campus Marketplace API
  version: 1.0.0
  description: 校园二手交易与闲置物品流转平台 API
servers:
  - url: http://localhost:7001
paths:
...
  /api/categories/{id}:
    get:
      operationId: getCategory
      summary: 获取分类详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 分类详情
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Category"
        "404":
          description: 分类不存在
  /api/items:
    get:
      operationId: listItems
      summary: 浏览商品列表
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: keyword
          in: query
          schema: { type: string }
        - name: sortBy
          in: query
          schema: { type: string, enum: [created_at, price] }
        - name: sortOrder
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        "200":
          description: 商品列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemPage"
    post:
      operationId: createItem
      summary: 发布商品
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemInput"
      responses:
        "201":
          description: 发布成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Item"
        "400":
          description: 输入无效
        "401":
          description: 未认证
  /api/items/{id}:
    get:
      operationId: getItem
      summary: 获取商品详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 商品详情
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Item"
        "404":
          description: 商品不存在
    put:
      operationId: updateItem
      summary: 更新商品
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemInput"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Item"
        "403":
          description: 无权操作
        "404":
          description: 商品不存在
    delete:
      operationId: deleteItem
      summary: 删除商品
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
        "403":
          description: 无权操作
        "404":
          description: 商品不存在
...
  /api/favorites:
    get:
      operationId: listFavorites
      summary: 获取我的收藏
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 收藏列表
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Favorite"
        "401":
          description: 未认证
  /api/favorites/{itemId}:
    post:
      operationId: addFavorite
      summary: 收藏商品
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: 收藏成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Favorite"
        "404":
          description: 商品不存在
    delete:
      operationId: removeFavorite
      summary: 取消收藏
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 取消成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
        "404":
          description: 收藏记录不存在
  /api/favorites/check/{itemId}:
    get:
      operationId: checkFavorite
      summary: 检查是否已收藏
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 检查结果
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [isFavorited]
                    properties:
                      isFavorited:
                        type: boolean
  /api/admin/items:
    get:
      operationId: adminListItems
      summary: 管理员获取所有商品
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_review, listed, reserved, sold, rejected, removed]
      responses:
        "200":
          description: 商品列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemPage"
        "403":
          description: 需要管理员权限
  /api/admin/items/{id}/approve:
    post:
      operationId: approveItem
      summary: 审核通过商品
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 审核通过
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Item"
        "400":
          description: 商品状态不允许审核
        "403":
          description: 需要管理员权限
  /api/admin/items/{id}/reject:
    post:
      operationId: rejectItem
      summary: 审核拒绝商品
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: 审核拒绝
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Item"
        "400":
          description: 商品状态不允许审核
        "403":
          description: 需要管理员权限
...
  /api/admin/categories:
    post:
      operationId: adminCreateCategory
      summary: 管理员创建分类
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Category"
        "403":
          description: 需要管理员权限
  /api/admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      summary: 管理员更新分类
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryInput"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Category"
        "403":
          description: 需要管理员权限
        "404":
          description: 分类不存在
    delete:
      operationId: adminDeleteCategory
      summary: 管理员删除分类
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
        "400":
          description: 分类下有子分类或商品
        "403":
          description: 需要管理员权限
        "404":
          description: 分类不存在
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Category:
      type: object
      required: [id, name, parentId, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: [integer, "null"]
        sortOrder:
          type: integer
...
    CreateCategoryInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        parentId:
          type: [integer, "null"]
        sortOrder:
          type: integer
    UpdateCategoryInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        sortOrder:
          type: integer
    Item:
      type: object
      required:
        [
          id,
          sellerId,
          categoryId,
          title,
          price,
          quantity,
          availableQuantity,
          status,
          images,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        categoryId:
          type: integer
        title:
          type: string
        description:
          type: [string, "null"]
        price:
          type: number
          minimum: 0
        quantity:
          type: integer
          minimum: 0
        availableQuantity:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [pending_review, listed, reserved, sold, rejected, removed]
        rejectReason:
          type: [string, "null"]
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        seller:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            studentId:
              type: string
        category:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
    CreateItemInput:
      type: object
      additionalProperties: false
      required: [categoryId, title, price]
      properties:
        categoryId:
          type: integer
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 2000
        price:
          type: number
          minimum: 0.01
        quantity:
          type: integer
          minimum: 1
          default: 1
        images:
          type: array
          items:
            type: string
    UpdateItemInput:
      type: object
      additionalProperties: false
      properties:
        categoryId:
          type: integer
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 2000
        price:
          type: number
          minimum: 0.01
        quantity:
          type: integer
          minimum: 1
        images:
          type: array
          items:
            type: string
    ItemPage:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
    Favorite:
      type: object
      required: [id, userId, itemId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        itemId:
          type: integer
        createdAt:
          type: string
          format: date-time
        item:
          $ref: "#/components/schemas/Item"